Fix buffer overflow
ROM sizes were always of size 2^n - 1, but we calculated an index
modulo 2^n, allowing accidental accesses to rom[2^n - 1], causing a
segfault. However, this modulo *must* be of a power of two to ensure
a uniform distribution of bucket indices.

Therefore, we change the meaning of the time and space factors to be
exactly 2^n, rather than 2^n - 1.